+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Added method to get all the selected facets - instance.getAllSelectedFacets()
 - Added a method to get the total count of selected facets - instance.getSelectedFacetsCount()
 - Added a method to get the total count of selected facet values - instance.getSelectedFacetValueCount()
+- Added wrapperTag options to products config
 
 
 ### Changed